update to use searchParams correctly
