+++ packages/@tinacms/toolkit/src/components/TinaUI.tsx
@@ -61,7 +61,7 @@ export const TinaUI: React.FC<TinaUIProps> = ({ children, position }) => {
             />
           )}
           <ActiveFieldIndicator />
-          <ReverseActiveFieldIndicator />
+          {/* <ReverseActiveFieldIndicator /> */}
         </div>
         {/* Dragging across the iframe causes mouse events to stop propagating so there's a laggy feeling without this */}
         <div className={`${resizingSidebar ? 'pointer-events-none' : ''}`}>
